Component AssistantAvatar to reduce boilerplate
